Extracted out AreaGraph subcomponents into separate files
