+++ web/src/components/areagraph.js
@@ -5,8 +5,13 @@ import { first, last } from 'lodash';
 
 import formatting from '../helpers/formatting';
 import { modeOrder, modeColor } from '../helpers/constants';
-import { getCo2Scale } from '../helpers/scales';
+import { detectHoveredDatapointIndex } from '../helpers/graph';
 import { prepareGraphData } from '../helpers/data';
+
+import ExchangeLinearGradients from './graph/exchangelineargradients';
+import HoverLine from './graph/hoverline';
+import ValueAxis from './graph/valueaxis';
+import TimeAxis from './graph/timeaxis';
 
 const d3 = Object.assign(
   {},
@@ -21,74 +26,65 @@ const X_AXIS_HEIGHT = 20;
 const Y_AXIS_WIDTH = 35;
 const Y_AXIS_PADDING = 4;
 
-const Axis = ({
-  className,
-  label,
-  scale,
-  renderLine,
-  renderTick,
-  textAnchor,
-  transform,
-}) => (
-  <g
-    className={className}
-    transform={transform}
-    fill="none"
-    fontSize="10"
-    fontFamily="sans-serif"
-    textAnchor={textAnchor}
-    style={{ pointerEvents: 'none' }}
-  >
-    {label && <text className="label" transform="translate(35, 80) rotate(-90)">{label}</text>}
-    <path className="domain" stroke="currentColor" d={renderLine(scale.range())} />
-    {scale.ticks(5).map(renderTick)}
-  </g>
-);
+// TODO: Consider merging this method with prepareGraphData.
+// This method is consumed by the AreaGraph component in such a way that it only gets called
+// if one of its input parameters changes. So recalculation doesn't happen if a user e.g. hovers
+// over the graph triggering a tooltip.
+const getGraphState = (currentTime, data, displayByEmissions, electricityMixMode, width, height) => {
+  if (!data || !data[0]) return {};
 
-const TimeAxis = React.memo(({ scale, height }) => {
-  const renderLine = range => `M${range[0] + 0.5},6V0.5H${range[1] + 0.5}V6`;
-  const renderTick = v => (
-    <g key={`tick-${v}`} className="tick" opacity={1} transform={`translate(${scale(v)},0)`}>
-      <line stroke="currentColor" y2="6" />
-      <text fill="currentColor" y="9" dy="0.71em">{moment(v).format('LT')}</text>
-    </g>
-  );
+  let maxTotalValue = d3.max(data, d => (
+    displayByEmissions
+      ? (d.totalCo2Production + d.totalCo2Import + d.totalCo2Discharge) / 1e6 / 60.0 // in tCO2eq/min
+      : (d.totalProduction + d.totalImport + d.totalDischarge) // in MW
+  ));
+  const format = formatting.scalePower(maxTotalValue);
+  const formattingFactor = !displayByEmissions ? format.formattingFactor : 1;
+  maxTotalValue /= formattingFactor;
 
-  return (
-    <Axis
-      className="x axis"
-      scale={scale}
-      renderLine={renderLine}
-      renderTick={renderTick}
-      textAnchor="middle"
-      transform={`translate(-1 ${height - X_AXIS_HEIGHT - 1})`}
-    />
-  );
-});
+  // Prepare graph data
+  const {
+    datetimes,
+    exchangeKeys,
+    graphData,
+  } = prepareGraphData(data, displayByEmissions, electricityMixMode, formattingFactor);
 
-const ValuesAxis = React.memo(({ scale, label, width }) => {
-  const renderLine = range => `M6,${range[0] + 0.5}H0.5V${range[1] + 0.5}H6`;
-  const renderTick = v => (
-    <g key={`tick-${v}`} className="tick" opacity={1} transform={`translate(0,${scale(v)})`}>
-      <line stroke="currentColor" x2="6" />
-      <text fill="currentColor" x="9" y="3" dx="0.32em">{v}</text>
-    </g>
-  );
+  // Prepare stack - order is defined here, from bottom to top
+  let stackKeys = modeOrder;
+  if (electricityMixMode === 'consumption') {
+    stackKeys = stackKeys.concat(exchangeKeys);
+  }
+  const stackedData = d3.stack()
+    .offset(d3.stackOffsetDiverging)
+    .keys(stackKeys)(graphData);
 
-  return (
-    <Axis
-      className="y axis"
-      label={label}
-      scale={scale}
-      renderLine={renderLine}
-      renderTick={renderTick}
-      textAnchor="start"
-      transform={`translate(${width - Y_AXIS_WIDTH - 1} -1)`}
-    />
-  );
-});
+  // Prepare axes and graph scales
+  const timeScale = d3.scaleTime()
+    .domain([first(datetimes), currentTime ? moment(currentTime).toDate() : last(datetimes)])
+    .range([0, width]);
+  const valuesScale = d3.scaleLinear()
+    .domain([0, maxTotalValue * 1.1])
+    .range([height, Y_AXIS_PADDING]);
+  const area = d3.area()
+    .x(d => timeScale(d.data.datetime))
+    .y0(d => valuesScale(d[0]))
+    .y1(d => valuesScale(d[1]))
+    .defined(d => Number.isFinite(d[1]));
 
-const Graph = React.memo(({
+  return {
+    area,
+    datetimes,
+    exchangeKeys,
+    format,
+    graphData,
+    timeScale,
+    valuesScale,
+    stackKeys,
+    stackedData,
+  };
+};
+
+const Layers = React.memo(({
   area,
   datetimes,
   displayByEmissions,
@@ -103,19 +99,6 @@ const Graph = React.memo(({
   layerMouseOutHandler,
   svgRef,
 }) => {
-  const detectPosition = (ev) => {
-    if (!datetimes.length) return null;
-    const dx = ev.pageX
-      ? (ev.pageX - svgRef.current.getBoundingClientRect().left)
-      : (d3.touches(this)[0][0]);
-    const datetime = timeScale.invert(dx);
-    // Find data point closest to
-    let i = d3.bisectLeft(datetimes, datetime);
-    if (i > 0 && datetime - datetimes[i - 1] < datetimes[i] - datetime) i -= 1;
-    if (i > datetimes.length - 1) i = datetimes.length - 1;
-    return i;
-  };
-
   // Mouse hover events
   let mouseOutTimeout;
   const handleLayerMouseMove = (ev, layer, ind) => {
@@ -124,7 +107,7 @@ const Graph = React.memo(({
       mouseOutTimeout = undefined;
     }
     setSelectedLayerIndex(ind);
-    const i = detectPosition(ev);
+    const i = detectHoveredDatapointIndex(ev, datetimes, timeScale, svgRef);
     if (layerMouseMoveHandler) {
       const position = { x: ev.clientX - 7, y: svgRef.current.getBoundingClientRect().top - 7 };
       layerMouseMoveHandler(stackKeys[ind], position, layer[i].data._countryData);
@@ -164,99 +147,12 @@ const Graph = React.memo(({
   );
 });
 
-const ExchangeLinearGradients = React.memo(({
-  colorBlindModeEnabled,
-  exchangeKeys,
-  graphData,
-  timeScale,
-}) => {
-  const x1 = timeScale.range()[0];
-  const x2 = timeScale.range()[1];
-  const co2ColorScale = getCo2Scale(colorBlindModeEnabled);
-  const stopOffset = datetime => `${(timeScale(datetime) - x1) / (x2 - x1) * 100.0}%`;
-  const stopColor = (countryData, key) => (countryData.exchangeCo2Intensities
-    ? co2ColorScale(countryData.exchangeCo2Intensities[key]) : 'darkgray');
-
-  return (
-    <React.Fragment>
-      {exchangeKeys.map(key => (
-        <linearGradient gradientUnits="userSpaceOnUse" id={`areagraph-exchange-${key}`} key={key} x1={x1} x2={x2}>
-          {graphData.map(d => (
-            <stop
-              key={d.datetime}
-              offset={stopOffset(d.datetime)}
-              stopColor={stopColor(d._countryData, key)}
-            />
-          ))}
-        </linearGradient>
-      ))}
-    </React.Fragment>
-  );
-});
-
-const getMaxTotalValue = (data, displayByEmissions) =>
-  d3.max(data, d => (
-    displayByEmissions
-      ? (d.totalCo2Production + d.totalCo2Import + d.totalCo2Discharge) / 1e6 / 60.0 // in tCO2eq/min
-      : (d.totalProduction + d.totalImport + d.totalDischarge) // in MW
-  ));
-
 const getCurrentTime = state =>
   state.application.customDate || (state.data.grid || {}).datetime;
 
 // Regular production mode or exchange fill as a fallback
 const fillColor = (key, displayByEmissions) =>
   modeColor[key] || (displayByEmissions ? 'darkgray' : `url(#areagraph-exchange-${key})`);
-
-const getGraphState = (currentTime, data, displayByEmissions, electricityMixMode, width, height) => {
-  if (!data || !data[0]) return {};
-
-  let maxTotalValue = getMaxTotalValue(data, displayByEmissions);
-  const format = formatting.scalePower(maxTotalValue);
-  const formattingFactor = !displayByEmissions ? format.formattingFactor : 1;
-  maxTotalValue /= formattingFactor;
-
-  // Prepare graph data
-  const {
-    datetimes,
-    exchangeKeys,
-    graphData,
-  } = prepareGraphData(data, displayByEmissions, electricityMixMode, formattingFactor);
-
-  // Prepare stack - order is defined here, from bottom to top
-  let stackKeys = modeOrder;
-  if (electricityMixMode === 'consumption') {
-    stackKeys = stackKeys.concat(exchangeKeys);
-  }
-  const stackedData = d3.stack()
-    .offset(d3.stackOffsetDiverging)
-    .keys(stackKeys)(graphData);
-
-  // Prepare axes and graph scales
-  const timeScale = d3.scaleTime()
-    .domain([first(datetimes), currentTime ? moment(currentTime).toDate() : last(datetimes)])
-    .range([0, width - Y_AXIS_WIDTH]);
-  const valuesScale = d3.scaleLinear()
-    .domain([0, maxTotalValue * 1.1])
-    .range([height - X_AXIS_HEIGHT, Y_AXIS_PADDING]);
-  const area = d3.area()
-    .x(d => timeScale(d.data.datetime))
-    .y0(d => valuesScale(d[0]))
-    .y1(d => valuesScale(d[1]))
-    .defined(d => Number.isFinite(d[1]));
-
-  return {
-    area,
-    datetimes,
-    exchangeKeys,
-    format,
-    graphData,
-    timeScale,
-    valuesScale,
-    stackKeys,
-    stackedData,
-  };
-};
 
 const mapStateToProps = (state, props) => ({
   colorBlindModeEnabled: state.application.colorBlindModeEnabled,
@@ -290,8 +186,10 @@ const AreaGraph = ({
   useEffect(() => {
     const updateDimensions = () => {
       if (ref.current) {
-        const { width, height } = ref.current.getBoundingClientRect();
-        setContainer({ width, height });
+        setContainer({
+          width: ref.current.getBoundingClientRect().width - Y_AXIS_WIDTH,
+          height: ref.current.getBoundingClientRect().height - X_AXIS_HEIGHT,
+        });
       }
     };
     // Initialize dimensions if they are not set yet
@@ -328,12 +226,12 @@ const AreaGraph = ({
         scale={timeScale}
         height={container.height}
       />
-      <ValuesAxis
+      <ValueAxis
         label={displayByEmissions ? 'tCO2eq/min' : format.unit}
         scale={valuesScale}
         width={container.width}
       />
-      <Graph
+      <Layers
         area={area}
         datetimes={datetimes}
         displayByEmissions={displayByEmissions}
@@ -348,35 +246,14 @@ const AreaGraph = ({
         layerMouseOutHandler={layerMouseOutHandler}
         svgRef={ref}
       />
-      {Number.isInteger(selectedIndex) && (
-        <line
-          className="vertical-line"
-          style={{
-            display: 'block',
-            pointerEvents: 'none',
-            shapeRendering: 'crispEdges',
-          }}
-          x1={timeScale(graphData[selectedIndex].datetime)}
-          x2={timeScale(graphData[selectedIndex].datetime)}
-          y1={valuesScale.range()[0]}
-          y2={valuesScale.range()[1]}
-        />
-      )}
-      {Number.isInteger(selectedIndex) && stackedData[selectedLayerIndex] && (
-        <circle
-          r="6"
-          style={{
-            display: 'block',
-            pointerEvents: 'none',
-            shapeRendering: 'crispEdges',
-            stroke: 'black',
-            strokeWidth: 1.5,
-            fill: fillColor(stackKeys[selectedLayerIndex], displayByEmissions),
-          }}
-          cx={timeScale(graphData[selectedIndex].datetime)}
-          cy={valuesScale(stackedData[selectedLayerIndex][selectedIndex][1])}
-        />
-      )}
+      <HoverLine
+        graphData={graphData}
+        layerData={stackedData[selectedLayerIndex]}
+        fill={fillColor(stackKeys[selectedLayerIndex], displayByEmissions)}
+        selectedIndex={selectedIndex}
+        valuesScale={valuesScale}
+        timeScale={timeScale}
+      />
       <ExchangeLinearGradients
         graphData={graphData}
         timeScale={timeScale}
